if store is closed, order cannot be placed
