+++ web/src/pages/UserPage.jsx
@@ -24,6 +24,8 @@ import OrderTab from "../components/UserPage/orderTab";
 import Sidebar from "../components/UserPage/sideBar";
 
 import { dummyRestaurants } from "../assets/dummyRestaurants.js";
+import { isRestaurantOpenToday } from "../utils/isRestaurantOpenToday.js";
+
 
 // ADDRESS to GEOLOCATION: OpenCage API
 async function geocodeAddress(address) {
@@ -61,56 +63,6 @@ function getDistanceInKm(lat1, lon1, lat2, lon2) {
 
   const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
   return (R * c).toFixed(2); // distance in km, rounded to 2 decimals
-}
-
-// RESTAURANT OPEN TIMES
-function isRestaurantOpenToday(hoursArray, now = new Date()) {
-  const days = [
-    "Sunday",
-    "Monday",
-    "Tuesday",
-    "Wednesday",
-    "Thursday",
-    "Friday",
-    "Saturday",
-  ];
-  const today = days[now.getDay()];
-
-  const todayEntry = hoursArray.find((entry) => entry[today]);
-  if (!todayEntry || !todayEntry[today]) return false;
-
-  const { Opening, Closing } = todayEntry[today];
-
-  if (
-    !Opening ||
-    !Closing ||
-    Opening.length !== 4 ||
-    Closing.length !== 4 ||
-    Opening === Closing
-  ) {
-    return false;
-  }
-
-  const openHour = parseInt(Opening.slice(0, 2), 10);
-  const openMinute = parseInt(Opening.slice(2), 10);
-  const closeHour = parseInt(Closing.slice(0, 2), 10);
-  const closeMinute = parseInt(Closing.slice(2), 10);
-
-  const openTime = new Date(now);
-  openTime.setHours(openHour, openMinute, 0, 0);
-
-  const closeTime = new Date(now);
-  closeTime.setHours(closeHour, closeMinute, 0, 0);
-
-  // Overnight handling
-  if (closeTime <= openTime) {
-    const closeTimeNextDay = new Date(closeTime);
-    closeTimeNextDay.setDate(closeTimeNextDay.getDate() + 1);
-    return now >= openTime || now <= closeTimeNextDay;
-  }
-
-  // Normal hours
-  return now >= openTime && now <= closeTime;
 }
 
 export default function UserPage() {
@@ -327,11 +279,12 @@ export default function UserPage() {
           const orderData = docSnap.data();
           if (orderData.userId === userData.id) {
             restaurantOrders.push({ ...orderData, orderId: docSnap.id });
+            /*
             console.log("Checking order:", {
               fromRestaurant: restaurant.storeName,
               orderUserId: orderData.userId,
               currentUserId: userData.id,
-            });
+            });*/
           }
         });
