+++ web/src/utils/isRestaurantOpenToday.js
@@ -1,0 +1,50 @@
+// RESTAURANT OPEN TIMES
+export function isRestaurantOpenToday(hoursArray, now = new Date()) {
+  const days = [
+    "Sunday",
+    "Monday",
+    "Tuesday",
+    "Wednesday",
+    "Thursday",
+    "Friday",
+    "Saturday",
+  ];
+  
+  const today = days[now.getDay()];
+
+  const todayEntry = hoursArray.find((entry) => entry[today]);
+  if (!todayEntry || !todayEntry[today]) return false;
+
+  const { Opening, Closing } = todayEntry[today];
+
+  if (
+    !Opening ||
+    !Closing ||
+    Opening.length !== 4 ||
+    Closing.length !== 4 ||
+    Opening === Closing
+  ) {
+    return false;
+  }
+
+  const openHour = parseInt(Opening.slice(0, 2), 10);
+  const openMinute = parseInt(Opening.slice(2), 10);
+  const closeHour = parseInt(Closing.slice(0, 2), 10);
+  const closeMinute = parseInt(Closing.slice(2), 10);
+
+  const openTime = new Date(now);
+  openTime.setHours(openHour, openMinute, 0, 0);
+
+  const closeTime = new Date(now);
+  closeTime.setHours(closeHour, closeMinute, 0, 0);
+
+  // Overnight handling
+  if (closeTime <= openTime) {
+    const closeTimeNextDay = new Date(closeTime);
+    closeTimeNextDay.setDate(closeTimeNextDay.getDate() + 1);
+    return now >= openTime || now <= closeTimeNextDay;
+  }
+
+  // Normal hours
+  return now >= openTime && now <= closeTime;
+}
